fixed the code and now the player can move down only

+++ code/qx_overworld.py
@@ -37,10 +37,10 @@ class Overworld:
         #path
         self.paths = {}
         for obj in tmx_map.get_layer_by_name("Paths"):
-          pos = [(int(p.x + tile_size/2),(p.y + tile_size/2)) for p in obj.points]
+          pos = [(int(p.x + tile_size/2),int(p.y + tile_size/2)) for p in obj.points]
           start = obj.properties["start"]
           end = obj.properties["end"]
-          self.paths[end] = {"pos":pos, "start":start}
+          self.paths[end] = {"pos" : pos, "start" : start}
 
         #nodes & player
         for obj in tmx_map.get_layer_by_name("Nodes"):
@@ -65,7 +65,7 @@ class Overworld:
           self.move("down")
 
     def move(self,direction):
-      path_key = int(self.current_node.paths[direction][0])
+      path_key = self.current_node.paths[direction][0]
       path_reverse = True if self.current_node.paths[direction][-1] == "r" else False
       path = self.paths[path_key]["pos"][:] if not path_reverse else self.paths[path_key]["pos"][::-1]
       self.icon.start_move(path)
